Parse arrays, input pivot
I added an implementation for TOML array parsing. Then I took a pause on parsing and started on the SPIR-V part.

I decided we should interpret spv binaries, not spvasm. This will speed up parsing (in execution, perhaps not in code development) and be more useful.

Therefore, with the input pivot to spv, I modified the example set.