Data foundations
Built out the idea of how data is held in the program (within the size of the bound). Data can be a variable, a function, a value (used for temporaries), or a type.

Note: I am again following the pattern of writing out the code then debugging later. The code compiles, but I do not guarantee correctness yet.