Successful build
Forward declarations are apparently not simpatico with C++ modules. I am not sure what was wrong with my previous code (perhaps it was a compiler error), but in any case, reverting a couple modules to headers-implementations fixed the build error.

Hopefully, at some future time, I can return type and value to modules. For now, it is more important to be able to build and make progress.