Reordered decorations
I discovered that the reason the struct members were not named was because decorations were applied to the type after the variable already fetched its copy. Since I didn't think it feasible to make Variable have a Type*, I reordered when decorations were applied. Now, decorations are collected and can be applied at the construction of the decorated object. I also considered an alternate approach where decorations are stored in the data of the thing which needs to be decorated, but that approach is destructive (the decorations are replaced by the thing itself). Destructive processes may cause problems for simultaneous invocations, so it is best to pre-emptively avoid them.