Parse interface load
Progress on the compute shader example, most notably, parses the program for OpVariables to load to the interface rather than expecting them to be given at the entry point (this is a luxury for simple shaders for recent SPIR-V versions, neither of which we can guarantee).

The compute shader can parse entirely. The next step is to compute results for some of the new instructions and then control flow executions.

I also fixed a couple of small bugs in TOML printing. Since the check functionality doesn't look at the exact syntax, it is a weak spot in my testing.