TOML -> YAML
After some investigation, I realized that I was abusing TOML's inline tables. I looked at how it is supposed to be done, and honestly, I was not a fan. Maybe in the future, we can support proper TOML, but I don't want it as the default. Therefore, I replaced it with YAML.

All tests are currently passing, although I am sure there are some small bugs to iron out. We should also make some unit tests for format outputs (since that isn't tested with the check feature). I manually checked for YAML and it looked good, but I am trying to think ahead.