Instruction printing (verbose mode)
Small change to implement printing of instructions executed. This is mainly meant as a debugging feature and will be vital for interactive mode.

I didn't want to print instructions the same way they appear in spvasm disassembly. First: I find it confusing how disassembly rearranges operands. Second, I don't like the formatting. Third, if the spvasm formatting is desired, the user can simply use spirv-dis. Since verbose mode is intended primarily for debugging, I made choices to be most beneficial for that case. One, raw ids are printed. Two, operands are printed in order. It should still be easy to find where variables are set by searching for the string "%id =", so that functionality is preserved.