Bug fixing and exec implementation
Fixed some bugs, implemented some new instruction executions, updated the README to note a limitation on recursion.

I included the current output for the two tests (which can now be run by the example runner). I suspect that the output for the second test is incorrect- that is something I will need to investigate in the next commit.