Format output tests
Format outputs are under-tested. Parsing is handled generously by the example runner, but generation (prior to this commit) was not tested automatically (instead only through visual inspection of outputs).

This testing should catch a few corner cases and should be expanded in the future. For the time, I focused on testing JSON since that is the least used format.

Also, update documentation.