Debugger progress
The debugger is nearly finished. There are still some bugs to work out
(especially with stepping in to functions) and features to test, but the
behavior should be implemented.

Remove spvasm files now that the debugger can show the spv's contents.

Fix a bug in the trie- include a new test to prevent regression.