Specialization Constants
Implement specialization constants, which are inputs provided at
compile-time. Since spirv-run doesn't compile anything, it needs to
handle spec constants like both variables and values. They also need a
static value (in this case, to be resolvable before the main function
begins). All this required a reordering of when inputs are read.