Split SPIR-V parsing and initialization
Previously, static instructions would be executed as they were parsed.
This was a simpler and faster approach, but it had some notable
drawbacks:
1) Part of the program could be executed before the program was
   determined to be invalid or unsupported.
2) Static execution may depend on specialization constants, and
   therefore, the inputs needed to be passed in during parsing. This is
   an improper combination of several distinct tasks and required
   tracking of inputs before determining program validity, which is
   argubably more important.
3) Division of execution and parsing is very important for when several
   different invocations or inputs may be used.

The new approach splits the two tasks into different functions. Also, a
new program loader has been created to store variables necessary only
for SPIR-V binary parsing. This is a better division of fields, and
prevents improper access of the binary after its destruction.

I also noticed that the usage of namespaces within the source was
inconsistent and confusing. A couple modules (specifically token
and program), included their titular classes within a namespace.
Originally, I did this because I suspected that the names Token and
Program may be reused. However, the whole concept of importing modules
makes this issue less likely- each module needs to be imported
individually rather than a single header include carrying with it a
whole slew of names. I have seen other projects get away with the header
approach without many namespaces, so I think I can do this here, if not
permanently, then at least temporarily.

For a time, I considered putting each module within a namespace
corresponding to the source folder (for example, frame.cxx is in
the folder spv, so Data could be within namespace Spv). However, I
decided this is too verbose and not necessary now. I reserve the right
to change my mind in the future, but until that point (which may never
come), I wanted to remove the inconsistency.