Data manager and views
Replace the data array with a data manager and views (one for each
frame). This allows for better stack handling at the cost of complexity.
The improved stack handling is necessary for implementing recursion and
multiple invocations (such as for compute shaders).

Include a new example to demonstrate working recursion.