Bug fixes and new instructions
Fixed a couple bugs (the most notable was with multiplying matrix by a
vector, which had the dimensions swapped- I needed to update the example
which used the logic. I believe it is correct now, but more follow-up
could be helpful) and added some new instructions both in SPIR-V proper
and the GLSL extension set.