Initial ray tracing/query implementation (#7)
Ray Tracing
The Khronos ray tracing extension has been implemented, excepting the execution of multiple shader stages.

The ray generation execution model currently returns a payload that indicates whether the ray intersected a primitive; it does not execute a ray tracing pipeline yet. The examples only check that the instructions work and are not examples that would be used in producing a proper ray traced scene. Examples include input/output in json format, but yaml is also supported. The current acceleration structure (AS) input format contains all acceleration structures that will be used. At the very top is the metadata with information on what each acceleration structure contains. The top-level acceleration structure (TLAS) is at the beginning of all AS and the bottom-level acceleration structures (BLAS) follow. The interpreter will take in the AS format and convert it into an object-based form where the acceleration structure is constructed bottom-up.
Currently, the acceleration structure is represented as an AccelerationStructureManager object that consist of AccelerationStructure objects that consists of Node objects. The AccelerationStructureManager contains the root AccelerationStructure (the TLAS) and the acceleration structure input formatted with the interpreter's data types, specifically Aggregate. Some of the added examples will contain scene.txt files that have Desmos links that provide a visual.

Ray Query
The Khronos ray query extension is in progress.

---------

Co-authored-by: Vu Du <vu1.du@sarc.samsung.com>
Co-authored-by: Vu Du <vu1.du@samsung.com>