Format compression, new example with variable ptrs
Make small changes to json and yaml printing to put up to four
primitives on a single line. This should shorten files with long arrays
dramatically.

Include a new example which utilizes the VariablePointers capability and
pointer comparison instructions. It also uses LocalInvocationId (note,
not index), which is now implemented.