Float printing
Implement float printing to fix a bug with checking the new multi-tesc
example. Also, implement the last GLSL instruction needed to execute
said test. Since it is fairly complicated, I didn't thoroughly check the
output to be correct, but having a test will be make future developers
aware of any changes to the test result.