Image Write and start Bitcast
Implementation for simple image writes (very similar to the simple image
reads earlier). Also, fix a couple small image processing bugs.

Start the work for implementing OpBitcast. The instruction can often be
used to convert rather diverse types- which is especially difficult
because the internal representation of these types in the interpreter is
not guaranteed to be identical to what SPIR-V typically expects (for
example, primitives are a union of the data (and with higher precision,
bits will need to be shuffled)).