RT refactor
There were some memory issues and overcomplexity with the old raytracing
setup. In this commit (which is not production ready and will not
build), I replaced the old acceleration structure front end with a new
one. I still need to connect the old functionality to the new front, at
least until I can rework that as well.