Improve default variable names
Improve the names for variables on the program interface, such as by
reverting to the builtin type or using the name of the element type.
Neither of these methods will prevent name collision, but name collision
wasn't entirely preventable before, so if nothing else, it doesn't make
the problem worse.

Update a few tests to use the new defaulted names. They are more
explanatory, which I generally think is better.