Fix OpMatrixTimesVector
Somewhere I got confused about whether matrices are column-major or
row-major (they should be column-major) and that caused me to implement
OpMatrixTimesVector incorrectly. Fix the implementation and update tests
which relied on the old (incorrect) version. Also, verified that the new
version is correct by comparing the result with an online tool.