Triangle hit attribute and image compares
Include a new ray tracing test which utilizes hit attributes from a
triangle hit (instead of a custom intersection shader which was
previously implemented). The new test also writes to an image, which now
has comparison logic implemented.

Restructure the raytracing test directories to avoid bloat and
redundancy in test names. I have considered doing a more systematic
grouping of other examples (such as by stage), but have not had a need
for it until now.