Work on new sampler image example
Started work on a new example which includes image samplers. Moved image
tests to their own directory (like what was done with raytracing), since
I anticipate more in the future.

Included a fix to YAML indentation and a small change which decreases
the indent count for arrays of structs (the change helps automatic
detection of indentation size and decreases indentation levels, which is
generally a good thing). Updated a couple YAML files in examples to
reflect the new change.