Align rt transform matrix with other matrices
OpenGL and SPIR-V use column-major matrices, meaning that if we have a
mat4x3, then there are 4 columns by 3 rows, and that a matrix is a
collection of column vectors. This can be confusing when displaying or
inputting matrices since these column vectors appear horizontal and can
be confused with rows.

The transformation matrix of an instance node in raytracing must be of
dimensions 4x3. Previously, input and output appeared transposed to make
the columns look more like columns. However, I think this inconsistency
is likely more harmful than helpful, so I have now aligned the
formatting for all matrices.