Function parameter fixes
Although glslang will only use Variables as function parameters, SPIR-V
makes no limitation to the parameter type. Futhermore, the argument
passed to the function is shallow copied, and if this argument is a
pointer (which is notably how variables are treated), then changes made
to the variable will be kept after the function exits.

Implement these constraints in the interpreter and include a new
example which demonstrates correct functionality. Also, fix multi-tesc,
which mistakenly had inout function parameters.