Set uninitialized values to dummy
Previously, unitialized values in the shader would be initialized to 0.
This was good because it prevented nondeterministic results, but it
could be improved. Set uninitialized float values to NaN and other types
to 0x1ABC2DEF to increase visibility (and thereby help the user notice
that the values have not been initialized).