Implement Atomic and Extended Arithmetic insts
Implement one of the atomic instructions (the logic for the others
should be similar) and some of the extended arirthmetic instructions.
Technically, all arithmetic instructions should be aware of underflow
and overflow, so I will need to apply the logic I wrote for the extended
arithmetic to the regular arithmetic instructions. However, this will
likely come with a performance cost.