Clarify which matrix expected for Instance Nodes
There was some ambiguity as to which transformation matrix is expected
for an instance node. The DirectX spec says that the object to world
transformation (the matrix which computes object coordinates) is
expected, but the ray tracing implementation will actually use the world
to object (which calculates new ray details to determine intersections
according to the object). This means that the inverse would be
calculated by the driver or the BVH builder.

On the other hand, I think that Vulkan expects world to object matrices
to be provided, which would ostensibly avoid the inverse calculation, at
least for implementations which don't require the rayquery backtracking.

There are pros and cons to both approaches, but generally I favor
providing the world to object transformations. This is a departure from
what the interpreter has done historically and thus required a static
inverse calculation for all matrices in examples.

In the future, I plan to mitigate some of the drawbacks of the
worldToObject approach with another "scene builder" tool to convert from
higher-level descriptions of the scene to BVH data which the interpreter
expects or OBJ files which can be rendered by 3D visualization software.