YAML compact mapping within sequence form
There is a special compact form which can be used in YAML when a mapping
(aka struct/object) is used within a sequence (aka array/list).

The previous default (expanded form):
a:
-
  b: 0
  c: 1
  d: 2
-
  b: 3
  c: 4
  d: 5
can be compacted to:
a:
- b: 0
  c: 1
  d: 2
- b: 3
  c: 4
  d: 5

Implement the code to support this new form in printing and parsing,
fixing bugs along the way and improving the testing coverage. The new
form is especially helpful for tesc shaders, which use gl_in.

Although the compact form is used by default, the previous form should
still be supported. As such, there are parsing unit tests as well as a
practical usage in example undef-legal.