Fix float printing
There were a couple small bugs in the float printing algorithm. I
decided a better approach to fixing in-place would be to redesign the
algorithm. I think the new approach is cleaner (it is also closer to my
original approach), and I added new test cases to prevent future
regressions.