Raytracing substage variable sharing (#20)
Shaders in a raytracing pipe may share variables placed by a layout
location/set. Previously, any such variables were passed by value from
the main stage (rgen) to any substages, however, it is possible for the
substages to modify these buffers, and the changes must persist after
the substage returns. This suggests a pass-by-reference approach rather
than copying to the substage.

Further, variables may be identified in a substage but not in the main
stage. These ought to be handled as if they had been present in the main
stage since they could be shared between several substages (and it is
easier to pass input uniformly to the main stage rather than using
fragmented inputs per shader stage). Unfortunately, it was not possible
to eliminate all shader-specific inputs due to the existence of shader
record inputs. Therefore, these are the only variables expected per
stage (as defined from the SBT). All other variables are copied to the
interface of the main stage and can therefore be provided as regular.