Fix inconsistency with coop matrix
The construction of a variable's value must be delayed because the
SPIR-V language allows for forward pointers. However, variables need
constructed values to name themselves and determine their type (which is
important in deciding whether they are threaded). Delay some of the
steps in program parsing to guarantee the value is constructed before it
is used.

This clarity resolves a crash with a particularly difficult cooperative
matrix example, where the matrix exists at static scope and is
initialized by element setting. I don't think this example will be
common in practice, but it is valid and needs to be accommodated for.